Add `--version` when testing if `bw` command works
The bw cli program recently changed so that running `bw` without any
arguments returned a failing status, causing `subprocess` to raise a
`CalledProcessError` that resulted in the lookup module
failing. Invoking the program with a no-op argument like `--version`
causes it to exit successfully, which allows us to test for its
presence.

Fixes #14.