chore[bw]: move cache to global variable
The `LookupModule` is initialised during every lookup. This results in a
new Bitwarden object for every lookup removing the benefit of the cache.

However, importing the bitwarden Lookup module is only done at specific
stages during a playbook and not run for every lookup itself. So
changing the cache to a global variable makes it persistent for as long
as possible and thus making the cache actually benefical for improving
the executing time.

Fixes #11