feat[bw]: make logged_in rely on global variable
During every lookup the Bitwarden object verifies whether it is logged
in. This executes CLI and HTTP calls which take time, which is
unnecessary for most of the time. Therefore I've implemented a global
variable which is undefined initially, thus ensuring that `logged_in`
function is executed and otherwise use the previous versions.

Manual verification shows that this decreases overall execution time of
lookup plugin